Refactor Pokemon Class
