+++ Pokemon/Pokemon.vcxproj
@@ -157,6 +157,7 @@
     <ClCompile Include="main.cpp" />
     <ClCompile Include="Player.cpp" />
     <ClCompile Include="Player.hpp" />
+    <ClCompile Include="Pokemon.cpp" />
     <ClCompile Include="Utility.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -168,6 +169,7 @@
   <ItemGroup>
     <ClInclude Include="EPokemonChoice.hpp" />
     <ClInclude Include="EPokemonType.hpp" />
+    <ClInclude Include="Pokemon.hpp" />
     <ClInclude Include="Utility.hpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
