+++ Pokemon/Pokemon/x64/Debug/Pokemon.log
@@ -1,39 +1,68 @@
 ﻿  main.cpp
+C:\Users\utkri\source\repos\Pokemon\Pokemon\EPokemonType.hpp(1,12): error C2011: 'PokemonType': 'enum' type redefinition
+  (compiling source file 'main.cpp')
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\EPokemonType.hpp(1,12):
+      see declaration of 'PokemonType'
+  
+C:\Users\utkri\source\repos\Pokemon\Pokemon\Pokemon.hpp(5,7): error C2011: 'Pokemon': 'class' type redefinition
+  (compiling source file 'main.cpp')
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\Pokemon.hpp(5,7):
+      see declaration of 'Pokemon'
+  
+C:\Users\utkri\source\repos\Pokemon\Pokemon\main.cpp(57,92): error C2027: use of undefined type 'Pokemon'
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\Pokemon.hpp(5,7):
+      see declaration of 'Pokemon'
+  
+C:\Users\utkri\source\repos\Pokemon\Pokemon\main.cpp(152,10): error C2079: 'pokemon' uses undefined class 'Pokemon'
+C:\Users\utkri\source\repos\Pokemon\Pokemon\main.cpp(152,42): error C2027: use of undefined type 'PokemonType'
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\EPokemonType.hpp(1,12):
+      see declaration of 'PokemonType'
+  
+C:\Users\utkri\source\repos\Pokemon\Pokemon\main.cpp(152,42): error C2065: 'Electric': undeclared identifier
+C:\Users\utkri\source\repos\Pokemon\Pokemon\main.cpp(152,17): error C2440: 'initializing': cannot convert from 'initializer list' to 'int'
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\main.cpp(152,17):
+      The initializer contains too many elements
+  
+C:\Users\utkri\source\repos\Pokemon\Pokemon\main.cpp(153,15): error C2665: 'Player::Player': no overloaded function could convert all the argument types
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.hpp(16,2):
+      could be 'Player::Player(std::string,Pokemon)'
+          C:\Users\utkri\source\repos\Pokemon\Pokemon\main.cpp(153,15):
+          'Player::Player(std::string,Pokemon)': cannot convert argument 2 from 'int' to 'Pokemon'
+              C:\Users\utkri\source\repos\Pokemon\Pokemon\main.cpp(153,27):
+              use of undefined type 'Pokemon'
+              C:\Users\utkri\source\repos\Pokemon\Pokemon\Pokemon.hpp(5,7):
+              see declaration of 'Pokemon'
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\main.cpp(153,15):
+      while trying to match the argument list '(const char [8], int)'
+  
   Player.cpp
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.hpp(8,10): error C3646: 'ChosenPokemon': unknown override specifier
+C:\Users\utkri\source\repos\Pokemon\Pokemon\EPokemonType.hpp(1,12): error C2011: 'PokemonType': 'enum' type redefinition
   (compiling source file 'Player.cpp')
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\EPokemonType.hpp(1,12):
+      see declaration of 'PokemonType'
   
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.hpp(8,2): error C4430: missing type specifier - int assumed. Note: C++ does not support default-int
-  (compiling source file 'Player.cpp')
-  
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.hpp(15,27): error C2061: syntax error: identifier 'Pokemon'
-  (compiling source file 'Player.cpp')
-  
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(8,9): error C2084: function 'Player::Player(void)' already has a body
-      C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.hpp(10,2):
+C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(9,9): error C2084: function 'Player::Player(void)' already has a body
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.hpp(11,2):
       see previous definition of 'Player::Player'
   
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(10,2): error C2065: 'ChosenPokemon': undeclared identifier
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(10,18): error C3861: 'Pokemon': identifier not found
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(13,34): error C2061: syntax error: identifier 'Pokemon'
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(16,2): error C2065: 'ChosenPokemon': undeclared identifier
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(16,18): error C2065: 'chosenPokemon': undeclared identifier
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(22,2): error C2065: 'ChosenPokemon': undeclared identifier
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(22,24): error C2039: 'ChosenPokemon': is not a member of 'Player'
-      C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.hpp(4,7):
-      see declaration of 'Player'
+C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(31,53): error C2027: use of undefined type 'PokemonType'
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\EPokemonType.hpp(1,12):
+      see declaration of 'PokemonType'
   
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(30,3): error C2065: 'ChosenPokemon': undeclared identifier
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(30,19): error C3861: 'Pokemon': identifier not found
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(33,3): error C2065: 'ChosenPokemon': undeclared identifier
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(33,19): error C3861: 'Pokemon': identifier not found
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(36,3): error C2065: 'ChosenPokemon': undeclared identifier
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(36,19): error C3861: 'Pokemon': identifier not found
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(39,3): error C2065: 'ChosenPokemon': undeclared identifier
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(39,19): error C3861: 'Pokemon': identifier not found
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(43,40): error C2065: 'ChosenPokemon': undeclared identifier
-  Player.hpp
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.hpp(8,10): error C3646: 'ChosenPokemon': unknown override specifier
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.hpp(8,2): error C4430: missing type specifier - int assumed. Note: C++ does not support default-int
-C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.hpp(15,27): error C2061: syntax error: identifier 'Pokemon'
+C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(31,53): error C2065: 'Grass': undeclared identifier
+C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(34,54): error C2027: use of undefined type 'PokemonType'
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\EPokemonType.hpp(1,12):
+      see declaration of 'PokemonType'
+  
+C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(34,54): error C2065: 'Fire': undeclared identifier
+C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(37,52): error C2027: use of undefined type 'PokemonType'
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\EPokemonType.hpp(1,12):
+      see declaration of 'PokemonType'
+  
+C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(37,52): error C2065: 'Water': undeclared identifier
+C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(40,51): error C2027: use of undefined type 'PokemonType'
+      C:\Users\utkri\source\repos\Pokemon\Pokemon\EPokemonType.hpp(1,12):
+      see declaration of 'PokemonType'
+  
+C:\Users\utkri\source\repos\Pokemon\Pokemon\Player.cpp(40,51): error C2065: 'Electric': undeclared identifier
   Generating Code...
