+++ Pokemon/main.cpp
@@ -2,93 +2,11 @@
 #include <string>
 #include <limits>
 #include "EPokemonChoice.hpp"
+#include "Player.hpp"
 #include "EPokemonType.hpp"
 #include "Utility.hpp"
+#include "Pokemon.hpp"
 using namespace std;
-
-class Pokemon
-{
-public:
-	string Name;
-	PokemonType Type;
-	int Health;
-
-	Pokemon() 
-	{
-		Name = "Pikachu";
-		Type = PokemonType::Electric;
-		Health = 100;
-	}
-
-	Pokemon(string name, PokemonType type, int health)
-	{
-		Name = name;
-		Type = type;
-		Health = health;
-	}
-
-	Pokemon(const Pokemon& other) 
-	{
-		Name = other.Name;
-		Type = other.Type;
-		Health = other.Health;
-	}
-
-	~Pokemon() 
-	{
-	}
-
-	void Attack()
-	{
-		cout << "Pokemon is attacking!" << endl;
-	}
-};
-
-#include "Player.hpp"
-
-//class Player
-//{
-//public:
-//	string Name;
-//	Pokemon ChosenPokemon;
-//
-//	Player() {
-//		Name = "Trainer";
-//		ChosenPokemon = Pokemon();
-//	}
-//
-//	Player(std::string name, Pokemon chosenPokemon) 
-//	{
-//		Name = name;
-//		ChosenPokemon = chosenPokemon;
-//	}
-//
-//	Player(const Player& other)
-//	{
-//		Name = other.Name;
-//		ChosenPokemon = other.ChosenPokemon;
-//	}
-//
-//	void ChoosePokemon(int choice)
-//	{
-//		switch (choice)
-//		{
-//		case 1:
-//			ChosenPokemon = Pokemon("Bulbasaur", PokemonType::Grass, 100);
-//			break;
-//		case 2:
-//			ChosenPokemon = Pokemon("Charmander", PokemonType::Fire, 100);
-//			break;
-//		case 3:
-//			ChosenPokemon = Pokemon("Squirtle", PokemonType::Water, 100);
-//			break;
-//		default:
-//			ChosenPokemon = Pokemon("Pikachu", PokemonType::Electric, 100);
-//			break;
-//		}
-//		cout << "System: You have chosen " << ChosenPokemon.Name << " as your first Pokemon!\n";
-//	}
-//};
 
 class ProfessorOak
 {
